Heart_Failure_Prediction.ipynb
This is the analysis of the Hear Failure Prediction Dataset in kaggle.com. Full link to the data source can be found below. The analysis was carried out as part of the Data Science with Python and R (4IZ565) coursework at VSE.

Source: https://www.kaggle.com/datasets/fedesoriano/heart-failure-prediction?resource=download